+++ src/pages/AuthPage/index.ts
@@ -1,0 +1,1 @@
+export * from './AuthPage';
